[frontend] fix couponItem link, remove unused page
[frontend] fix build error
