--- frontend/src/pages/coupon/[couponID].tsx
@@ -1,5 +1,0 @@
-const EachCoupon = () => {
-  return <></>;
-};
-
-export default EachCoupon;
